fix: widen certification hours to float, filter malformed proposals
- Certification.hours was int-only, rejecting genuine half-hour course
  durations (4.5h, 7.5h) found in the user's actual updated résumés.
- proposer.attach_evidence now dry-runs every update_field/new_item
  through the same validation store.py applies at write time, dropping
  anything that wouldn't actually apply (wrong shape, missing required
  field, unresolvable path) instead of surfacing model garbage for a
  human to reject by hand. Found by running the indexer against real,
  freshly-edited résumé files: ~40 of 49 proposed changes were malformed
  duplicates/hallucinations for the same two or three real facts.

+++ jarvis/resume/schema.py
@@ -127,7 +127,11 @@ class Project(StrictModel):
 
 class Certification(StrictModel):
     name: str
-    hours: Optional[int] = None
+    # float, not int: real course certificates commonly report half-hour
+    # durations (4.5h, 7.5h) -- an int-only field would reject genuine,
+    # correctly-extracted data. Existing whole-number values still validate
+    # fine, so this is a pure widening, no migration needed.
+    hours: Optional[float] = None
     status: Literal["in_progress", "completed"] = "completed"
     date: Optional[str] = None  # completed date, or expected date if in_progress
     evidence: list[Evidence] = Field(default_factory=list)
